cleans up unnecessary code, follows good code style
