+++ starts-plugin/src/it/static-nested-it/src/main/java/first/First.java
@@ -1,0 +1,17 @@
+package first;
+
+public class First {
+
+    public First() { }
+
+   static class Nested {
+       static Short number = new Short((short)0);
+
+        Nested() { }
+
+       public static String getNestedNumberAsString() {
+           return number.toString();
+       }
+    }
+
+}
